refactor(analytics): break tracker↔event_properties import cycle
